requirements.txt
python-telegram-bot
python-dotenv
requests
google-generativeai
psycopg2-binary
Flask[async]
cryptography
pytest
gunicorn
uvicorn
fastapi
matplotlib
